+++ src/lua-wrap/render-table/particle-lua.cpp
@@ -109,7 +109,7 @@ namespace mbm
         */
         PARTICLE *  particle = getParticleFromRawTable(lua, 1, 1);
         const char *what     = luaL_checkstring(lua, 2);
-        const int   len      = strlen(what);
+        const int   len      = static_cast<int>(strlen(what));
         switch (len)
         {
             case 1:
@@ -250,7 +250,7 @@ namespace mbm
         */
         PARTICLE *  particle = getParticleFromRawTable(lua, 1, 1);
         const char *what     = luaL_checkstring(lua, 2);
-        const int   len      = strlen(what);
+        const int   len      = static_cast<int>(strlen(what));
         switch (len)
         {
             case 1:
