+++ src/lua-wrap/vec3-lua.cpp
@@ -521,7 +521,7 @@ namespace mbm
         */
         const char *what = luaL_checkstring(lua, 2);
         VEC3 *      vec3 = getVec3FromRawTable(lua, 1, 1);
-        const int   len  = strlen(what);
+        const int   len  = static_cast<int>(strlen(what));
         if (len == 1)
         {
             switch (what[0])
@@ -550,7 +550,7 @@ namespace mbm
         */
         const char *what = luaL_checkstring(lua, 2);
         VEC3 *      vec3 = getVec3FromRawTable(lua, 1, 1);
-        const int   len  = strlen(what);
+        const int   len  = static_cast<int>(strlen(what));
         if (len == 1)
         {
             switch (what[0])
